Move scheduler into tasks/fitbitpoll
